requirements.txt
#streamlit
gTTS==2.2.2
googletrans==3.1.0a0
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
openai==0.28
tiktoken
PyPDF2==3.0.1
keras
numpy
opencv-python-headless
Pillow
tensorflow==2.12
